fix some compiler warnings
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@3025 8aca7d54-2c30-db11-9de9-000461428c89

+++ besch/groundobj_besch.h
@@ -58,7 +58,7 @@ public:
 		return static_cast<const bildliste2d_besch_t *>(get_child(2))->get_anzahl();
 	}
 
-	int get_seasons() const { return number_of_seasons; }
+	uint8 get_seasons() const { return number_of_seasons; }
 
 	uint32 get_speed() const { return speed; }
